EDA file added
EDA Case Study
Loan Risk Analysis
The Fintech companies and banks earning source is from the interests earned from the
loans so it is essential to lend loans to customers. But it is also essential to lend loans to
only customers who are likely to repay the loans.
We analyzed the criteria on which the loans should be distributed among defaulters and
Non defaulters.